Learning portal/requirements.txt
bootstrap4==0.1.0
branca==0.4.1
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cssutils==1.0.2
email-to==0.1.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
folium==0.11.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
premailer==3.7.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytube3==9.6.4
pytz==2020.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
